backend/requirements.txt
fastapi
python-dotenv
requests
uvicorn[standard]
python-multipart
pymupdf4llm
tiktoken